refactor(component): rename component name to LightboxMediaViewer
