+++ packages/react/.env.example
@@ -29,5 +29,5 @@ GEO_API=<endpoint for getting user country by geolocation>
 # Carbon Expressive Theme
 REACT_STORYBOOK_CARBON_EXPRESSIVE=<Boolean to turn on/off the carbon expressive theme in storybook>
 
-# Lightbox
-DDS_LIGHTBOX=<Boolean to turn on/off the lightbox component in storybook>
+# Lightbox Media Viewer
+DDS_LIGHTBOX_MEDIA_VIEWER=<Boolean to turn on/off the LightboxMediaViewer component in storybook>
