+++ packages/react/src/components/LightboxMediaViewer/__stories__/index.scss
@@ -5,4 +5,4 @@
 // LICENSE file in the root directory of this source tree.
 //
 @import '@carbon/grid/scss/grid.scss';
-@import '../../../../../styles/scss/components/lightbox/_lightbox.scss';
+@import '../../../../../styles/scss/components/lightbox-media-viewer/_lightbox-media-viewer.scss';
