+++ packages/styles/scss/ibm-dotcom-styles.scss
@@ -17,7 +17,7 @@
 @import 'components/footer/index';
 @import 'components/horizontalrule/horizontalrule';
 @import 'components/layout/layout';
-@import 'components/lightbox/lightbox';
+@import 'components/lightbox-media-viewer/lightbox-media-viewer';
 @import 'components/link-with-icon/link-with-icon';
 @import 'components/masthead/index';
 @import 'components/tableofcontents/tableofcontents';
